+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Selection of an entity from the connection matrix (context menu)
+- [Mac Address field in entities list](https://github.com/christophe-calmejane/Hive/issues/38)
+- [ClockDomain locked state in entities list](https://github.com/christophe-calmejane/Hive/issues/133)
 
 ### Fixed
 - Current selection in list of entities lost when changing the filter